Switch a whole bunch of shared_ptrs to raw pointers
This was difficult to split up into smaller commits, hence the huge dump
of changes.

I originally wanted to go big on "modern C++" with this project, which
is why I ended up with so many shared_ptrs: I wanted no raw pointers and
thought that was the way things were done now.

I tried it - it's fine for this project - but I prefer using unique_ptrs
and giving out raw pointers when users aren't expected to take
ownership.